phase_rs

Module combinator

The combinator version of the language.

Example

To parse from string src, typecheck, evaluate, and print a command, the following can be run:

use phase_rs::combinator::{
   parsing::command,
   syntax::{ToDoc, normal::TermN},
};
use winnow::{LocatingSlice, Parser, ascii::multispace0, combinator::terminated};
fn parse_and_eval(src: &str) {
  let parsed = terminated(command, multispace0)
      .parse(LocatingSlice::new(src)).unwrap();
  let (_env, checked) = parsed.check().unwrap();
  let mut evalled: TermN = checked.eval();
  evalled.squash();
  let quoted = evalled.quote();
  let raw = quoted.to_raw();
  println!("Evaluated:\n{}\n", raw.to_doc().pretty(60));
}

Modules

command

A syntax of commands and their parsing.

eval

Simple term evaluation, expanding and simplifying terms.

eval_circ

Evaluation to circuit normal form, for extracting quantum circuits.

parsing

Parsing functions for raw syntax.

syntax

Syntax definitions for raw, typed, normal, and circuit-normal syntax.

typecheck

Functions and datastructures for type checking

unitary

Contains functions for converting normal-form syntax to matrices.