phase_rs::combinator::typed_syntax

Enum TermT

pub enum TermT {
    Comp(Vec<TermT>),
    Tensor(Vec<TermT>),
    Id(TermType),
    Phase(Phase),
    IfLet {
        pattern: PatternT,
        inner: Box<TermT>,
    },
    Gate {
        name: Name,
        def: Box<TermT>,
    },
    Inverse(Box<TermT>),
    Sqrt(Box<TermT>),
}

Syntax of typed terms

Variants

Comp(Vec<TermT>)

A non-empty composition โt_1 ; โฆ ; t_nโ

Tensor(Vec<TermT>)

A tensor โt_1 x โฆ x t_nโ

Id(TermType)

An identity โid(n)โ

Phase(Phase)

A (global) phase operator, e.g. โ-1โ or โph(0.1pi)โ

IfLet

An โif letโ statement, โif let pattern then innerโ

Fields

pattern: PatternT

Pattern to match on in โif letโ

inner: Box<TermT>

Body of the โif letโ

Gate

Top level symbol, a named gate

Fields

name: Name

Name of symbol/gate

def: Box<TermT>

Definition of symbol

Inverse(Box<TermT>)

Inverse of a term โt ^ -1โ

Sqrt(Box<TermT>)

Square root of a term โsqrt(t)โ

Implementations

impl TermT

pub fn get_type(&self) -> TermType

Returns the type of this term

impl TermT

pub fn eval<B: Buildable>(&self) -> B

Evaluate a term to a given Buildable type, expanding top level definitions and evaluating inverse and sqrt macros. In particular this can be used to generate a TermN from a TermT.

pub fn eval_circ(&self) -> TermC

Returns a TermC representing the โcircuit-normal-formโ of the term.

impl TermT

pub fn to_raw(&self) -> TermR<()>

Convert to a raw term.