phase_rs::normal_syntax

Trait Buildable

pub trait Buildable {
    // Required methods
    fn comp(iter: impl DoubleEndedIterator<Item = Self>, ty: &TermType) -> Self;
    fn tensor(iter: impl Iterator<Item = Self>) -> Self;
    fn atom(atom: AtomN) -> Self;
}

Trait for objects that can built with compositions, tensors, or from an AtomN.

Required Methods

fn comp(iter: impl DoubleEndedIterator<Item = Self>, ty: &TermType) -> Self

Build a composition object from a sequence of subobjects and a given type. Subobjects should be given in diagrammatic order, not function composition order.

fn tensor(iter: impl Iterator<Item = Self>) -> Self

Build a tensor product from a sequence of subobjects.

fn atom(atom: AtomN) -> Self

Build an object from an atom.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.