phase_rs::raw_syntax::term

Type Alias TermR

pub type TermR<S> = Spanned<S, TermRInner<S>>;

Raw syntax term with text span. Represents a list of tensored terms composed together.

Aliased Type

pub struct TermR<S> {
    pub inner: TermRInner<S>,
    pub span: S,
}

Fields
inner: TermRInner<S>

Wrapped data

span: S

Text span

Implementations

impl<S: Clone> TermR<S>

pub fn check( &self, env: &Env, check_sqrt: Option<&S>, ) -> Result<TermT, TypeCheckError<S>>

Typecheck a raw term in given environment If check_sqrt is not None, then checks that the term is “composition free”